--- Sports2D/Utilities/common.py
@@ -1,871 +1,0 @@
-#!/usr/bin/env python
-# -*- coding: utf-8 -*-
-
-
-'''
-    ##################################################
-    ## Common classes and functions                 ##
-    ##################################################
-    
-    - A class for displaying several matplotlib figures in tabs.
-    - A function for interpolating sequences with missing data. 
-    It does not interpolate sequences of more than N contiguous missing data.
-
-'''
-
-
-## INIT
-import os
-import re
-import sys
-import cv2
-import subprocess
-import itertools as it
-import logging
-from datetime import datetime
-from pathlib import Path
-import argparse
-
-import numpy as np
-from scipy import interpolate
-import imageio_ffmpeg as ffmpeg
-from rtmlib import PoseTracker, BodyWithFeet, Wholebody, Body
-
-from PyQt5.QtWidgets import QMainWindow, QApplication, QWidget, QTabWidget, QVBoxLayout
-from matplotlib.backends.backend_qt5agg import FigureCanvasQTAgg as FigureCanvas
-from matplotlib.backends.backend_qt5agg import NavigationToolbar2QT as NavigationToolbar
-
-
-## AUTHORSHIP INFORMATION
-__author__ = "David Pagnon"
-__copyright__ = "Copyright 2023, Sports2D"
-__credits__ = ["David Pagnon"]
-__license__ = "BSD 3-Clause License"
-__version__ = "0.4.0"
-__maintainer__ = "David Pagnon"
-__email__ = "contact@david-pagnon.com"
-__status__ = "Development"
-
-
-## CLASSES
-class plotWindow():
-    '''
-    Display several figures in tabs
-    Taken from https://github.com/superjax/plotWindow/blob/master/plotWindow.py
-
-    USAGE:
-    pw = plotWindow()
-    f = plt.figure()
-    plt.plot(x1, y1)
-    pw.addPlot("1", f)
-    f = plt.figure()
-    plt.plot(x2, y2)
-    pw.addPlot("2", f)
-    '''
-    def __init__(self, parent=None):
-        self.app = QApplication.instance()
-        if not self.app:
-            self.app = QApplication(sys.argv)
-        self.MainWindow = QMainWindow()
-        self.MainWindow.setWindowTitle("Multitabs figure")
-        self.canvases = []
-        self.figure_handles = []
-        self.toolbar_handles = []
-        self.tab_handles = []
-        self.current_window = -1
-        self.tabs = QTabWidget()
-        self.MainWindow.setCentralWidget(self.tabs)
-        self.MainWindow.resize(1280, 720)
-        self.MainWindow.show()
-
-    def addPlot(self, title, figure):
-        new_tab = QWidget()
-        layout = QVBoxLayout()
-        new_tab.setLayout(layout)
-
-        figure.subplots_adjust(left=0.1, right=0.99, bottom=0.1, top=0.91, wspace=0.2, hspace=0.2)
-        new_canvas = FigureCanvas(figure)
-        new_toolbar = NavigationToolbar(new_canvas, new_tab)
-
-        layout.addWidget(new_canvas)
-        layout.addWidget(new_toolbar)
-        self.tabs.addTab(new_tab, title)
-
-        self.toolbar_handles.append(new_toolbar)
-        self.canvases.append(new_canvas)
-        self.figure_handles.append(figure)
-        self.tab_handles.append(new_tab)
-
-    def show(self):
-        self.app.exec_()
-        
-## FUNCTIONS
-def interpolate_zeros_nans(col, *args):
-    '''
-    Interpolate missing points (of value zero),
-    unless more than N contiguous values are missing.
-
-    INPUTS:
-    - col: pandas column of coordinates
-    - args[0] = N: max number of contiguous bad values, above which they won't be interpolated
-    - args[1] = kind: 'linear', 'slinear', 'quadratic', 'cubic'. Default: 'cubic'
-
-    OUTPUT:
-    - col_interp: interpolated pandas column
-    '''
-
-    if len(args)==2:
-        N, kind = args
-    if len(args)==1:
-        N = np.inf
-        kind = args[0]
-    if not args:
-        N = np.inf
-    
-    # Interpolate nans
-    mask = ~(np.isnan(col) | col.eq(0)) # true where nans or zeros
-    idx_good = np.where(mask)[0]
-    if len(idx_good) <= 4:
-        return col
-    
-    if 'kind' not in locals(): # 'linear', 'slinear', 'quadratic', 'cubic'
-        f_interp = interpolate.interp1d(idx_good, col[idx_good], kind="linear", bounds_error=False)
-    else:
-        f_interp = interpolate.interp1d(idx_good, col[idx_good], kind=kind, fill_value='extrapolate', bounds_error=False)
-    col_interp = np.where(mask, col, f_interp(col.index)) #replace at false index with interpolated values
-    
-    # Reintroduce nans if length of sequence > N
-    idx_notgood = np.where(~mask)[0]
-    gaps = np.where(np.diff(idx_notgood) > 1)[0] + 1 # where the indices of true are not contiguous
-    sequences = np.split(idx_notgood, gaps)
-    if sequences[0].size>0:
-        for seq in sequences:
-            if len(seq) > N: # values to exclude from interpolation are set to false when they are too long 
-                col_interp[seq] = np.nan
-    
-    return col_interp
-
-
-def natural_sort_key(s):
-    return [int(c) if c.isdigit() else c.lower() for c in re.split(r'(\d+)', s)]
-
-
-def make_homogeneous(list_of_arrays):
-    '''
-    Make a list of arrays (or a list of lists) homogeneous by padding with nans
-
-    Example: foo = [[array([nan, 656.02643776]), array([nan, nan])],
-                    [array([1, 2, 3]), array([1, 2])]]
-    becomes foo_updated = array([[[nan, 656.02643776, nan], [nan, nan, nan]],
-                                [[1., 2., 3.], [1., 2., nan]]])
-    Or foo = [[1, 2, 3], [1, 2], [3, 4, 5]]
-    becomes foo_updated = array([[1., 2., 3.], [1., 2., nan], [3., 4., 5.]])
-
-    INPUTS:
-    - list_of_arrays: list of arrays or list of lists
-
-    OUTPUT:
-    - np.array(list_of_arrays): numpy array of padded arrays
-    '''
-    
-    def get_max_shape(list_of_arrays):
-        if isinstance(list_of_arrays[0], list):
-            # Maximum length at the current level plus the max shape at the next level
-            return [max(len(arr) for arr in list_of_arrays)] + get_max_shape(
-                [item for sublist in list_of_arrays for item in sublist])
-        else:
-            # Determine the maximum shape across all list_of_arrays at this level
-            return [len(list_of_arrays)] + [max(arr.shape[i] for arr in list_of_arrays) for i in range(list_of_arrays[0].ndim)]
-
-    def pad_with_nans(list_of_arrays, target_shape):
-        '''
-        Recursively pad list_of_arrays with nans to match the target shape.
-        '''
-        if isinstance(list_of_arrays, np.ndarray):
-            # Pad the current array to the target shape
-            pad_width = [(0, max_dim - curr_dim) for curr_dim, max_dim in zip(list_of_arrays.shape, target_shape)]
-            return np.pad(list_of_arrays.astype(float), pad_width, constant_values=np.nan)
-        # Recursively pad each array in the list
-        return [pad_with_nans(array, target_shape[1:]) for array in list_of_arrays]
-
-    # Pad all missing dimensions of arrays with nans
-    list_of_arrays = [np.array(arr, dtype=float) if not isinstance(arr, np.ndarray) else arr for arr in list_of_arrays]
-    max_shape = get_max_shape(list_of_arrays)
-    list_of_arrays = pad_with_nans(list_of_arrays, max_shape)
-
-    return np.array(list_of_arrays)
-
-
-def resample_video(vid_output_path, fps, desired_framerate):
-    '''
-    Resample video to the desired fps using ffmpeg.
-    '''
-   
-    ffmpeg_path = ffmpeg.get_ffmpeg_exe()
-    new_vid_path = vid_output_path.parent / Path(vid_output_path.stem+'_2'+vid_output_path.suffix)
-    subprocess.run([ffmpeg_path, '-i', vid_output_path, '-filter:v', f'setpts={fps/desired_framerate}*PTS', '-r', str(desired_framerate), new_vid_path])
-    vid_output_path.unlink()
-    new_vid_path.rename(vid_output_path)
-
-
-def points2D_to_angles(points_list):
-    '''
-    If len(points_list)==2, computes clockwise angle of ab vector w.r.t. horizontal (e.g. RBigToe, RHeel) 
-    If len(points_list)==3, computes clockwise angle from a to c around b (e.g. Neck, Hip, Knee) 
-    If len(points_list)==4, computes clockwise angle between vectors ab and cd (e.g. Neck Hip, RKnee RHip)
-    
-    If parameters are float, returns a float between 0.0 and 360.0
-    If parameters are arrays, returns an array of floats between 0.0 and 360.0
-    '''
-
-    if len(points_list) < 2: # if not enough points, return None
-        return np.nan
-    
-    ax, ay = points_list[0]
-    bx, by = points_list[1]
-
-    if len(points_list)==2:
-        ux, uy = ax-bx, ay-by
-        vx, vy = 1,0
-    if len(points_list)==3:
-        cx, cy = points_list[2]
-        ux, uy = ax-bx, ay-by
-        vx, vy = cx-bx, cy-by
-
-    if len(points_list)==4:
-        cx, cy = points_list[2]
-        dx, dy = points_list[3]
-        ux, uy = bx-ax, by-ay
-        vx, vy = dx-cx, dy-cy
-
-    ang = np.arctan2(uy, ux) - np.arctan2(vy, vx)
-    ang_deg = np.degrees(ang)
-    # ang_deg = np.array(np.degrees(np.unwrap(ang*2)/2))
-    
-    return ang_deg
-
-
-def euclidean_distance(q1, q2):
-    '''
-    Euclidean distance between 2 points (N-dim).
-
-    INPUTS:
-    - q1: list of N_dimensional coordinates of point
-    - q2: idem
-
-    OUTPUTS:
-    - euc_dist: float. Euclidian distance between q1 and q2
-    '''
-
-    q1 = np.array(q1)
-    q2 = np.array(q2)
-    dist = q2 - q1
-
-    euc_dist = np.sqrt(np.sum( [d**2 for d in dist]))
-
-    return euc_dist
-
-def setup_pose_tracker(det_frequency, mode, pose_model = "HALPE_26"):
-    '''
-    Set up the RTMLib pose tracker with the appropriate model and backend.
-    If CUDA is available, use it with ONNXRuntime backend; else use CPU with openvino
-
-    INPUTS:
-    - det_frequency: int. The frequency of pose detection (every N frames)
-    - mode: str. The mode of the pose tracker ('lightweight', 'balanced', 'performance')
-    - tracking: bool. Whether to track persons across frames with RTMlib tracker
-
-    OUTPUTS:
-    - pose_tracker: PoseTracker. The initialized pose tracker object    
-    '''
-
-    # If CUDA is available, use it with ONNXRuntime backend; else use CPU with openvino
-    try:
-        import torch
-        import onnxruntime as ort
-        if torch.cuda.is_available() == True and 'CUDAExecutionProvider' in ort.get_available_providers():
-            device = 'cuda'
-            backend = 'onnxruntime'
-            logging.info(f"\nValid CUDA installation found: using ONNXRuntime backend with GPU.")
-        elif torch.cuda.is_available() == True and 'ROCMExecutionProvider' in ort.get_available_providers():
-            device = 'rocm'
-            backend = 'onnxruntime'
-            logging.info(f"\nValid ROCM installation found: using ONNXRuntime backend with GPU.")
-        else:
-            raise 
-    except:
-        try:
-            import onnxruntime as ort
-            if 'MPSExecutionProvider' in ort.get_available_providers() or 'CoreMLExecutionProvider' in ort.get_available_providers():
-                device = 'mps'
-                backend = 'onnxruntime'
-                logging.info(f"\nValid MPS installation found: using ONNXRuntime backend with GPU.")
-            else:
-                raise
-        except:
-            device = 'cpu'
-            backend = 'openvino'
-            logging.info(f"\nNo valid CUDA installation found: using OpenVINO backend with CPU.")
-
-    if det_frequency>1:
-        logging.info(f'Inference run only every {det_frequency} frames. Inbetween, pose estimation tracks previously detected points.')
-    elif det_frequency==1:
-        logging.info(f'Inference run on every single frame.')
-    else:
-        raise ValueError(f"Invalid det_frequency: {det_frequency}. Must be an integer greater or equal to 1.")
-
-    # Select the appropriate model based on the model_type
-    if pose_model.upper() == 'HALPE_26':
-        ModelClass = BodyWithFeet
-        logging.info(f"Using HALPE_26 model (body and feet) for pose estimation.")
-    elif pose_model.upper() == 'COCO_133':
-        ModelClass = Wholebody
-        logging.info(f"Using COCO_133 model (body, feet, hands, and face) for pose estimation.")
-    elif pose_model.upper() == 'COCO_17':
-        ModelClass = Body # 26 keypoints(halpe26)
-        logging.info(f"Using COCO_17 model (body) for pose estimation.")
-    else:
-        raise ValueError(f"Invalid model_type: {pose_model}. Must be 'HALPE_26', 'COCO_133', or 'COCO_17'. Use another network (MMPose, DeepLabCut, OpenPose, AlphaPose, BlazePose...) and convert the output files if you need another model. See documentation.")
-    logging.info(f'Mode: {mode}.\n')
-
-    # Initialize the pose tracker with Halpe26 model
-    pose_tracker = PoseTracker(
-        ModelClass,
-        det_frequency=det_frequency,
-        mode=mode,
-        backend=backend,
-        device=device,
-        tracking=False,
-        to_openpose=False)
-        
-    return pose_tracker
-
-def setup_webcam(webcam_id, save_video, vid_output_path, input_size):
-    '''
-    Set up webcam capture with OpenCV.
-
-    INPUTS:
-    - webcam_id: int. The ID of the webcam to capture from
-    - input_size: tuple. The size of the input frame (width, height)
-
-    OUTPUTS:
-    - cap: cv2.VideoCapture. The webcam capture object
-    - out_vid: cv2.VideoWriter. The video writer object
-    - cam_width: int. The actual width of the webcam frame
-    - cam_height: int. The actual height of the webcam frame
-    - fps: int. The frame rate of the webcam
-    '''
-
-    #, cv2.CAP_DSHOW launches faster but only works for windows and esc key does not work
-    cap = cv2.VideoCapture(webcam_id)
-    if not cap.isOpened():
-        raise ValueError(f"Error: Could not open webcam #{webcam_id}. Make sure that your webcam is available and has the right 'webcam_id' (check in your Config.toml file).")
-
-    # set width and height to closest available for the webcam
-    cap.set(cv2.CAP_PROP_FRAME_WIDTH, input_size[0])
-    cap.set(cv2.CAP_PROP_FRAME_HEIGHT, input_size[1])
-    cam_width = int(cap.get(cv2.CAP_PROP_FRAME_WIDTH))
-    cam_height = int(cap.get(cv2.CAP_PROP_FRAME_HEIGHT))
-    fps = cap.get(cv2.CAP_PROP_FPS)
-    if fps == 0: fps = 30
-    
-    if cam_width != input_size[0] or cam_height != input_size[1]:
-        logging.warning(f"Warning: Your webcam does not support {input_size[0]}x{input_size[1]} resolution. Resolution set to the closest supported one: {cam_width}x{cam_height}.")
-    
-    out_vid = None
-    if save_video:
-        # fourcc MJPG produces very large files but is faster. If it is too slow, consider using it and then converting the video to h264
-        # try:
-        #     fourcc = cv2.VideoWriter_fourcc(*'avc1') # =h264. better compression and quality but may fail on some systems
-        #     out_vid = cv2.VideoWriter(vid_output_path, fourcc, fps, (cam_width, cam_height))
-        #     if not out_vid.isOpened():
-        #         raise ValueError("Failed to open video writer with 'avc1' (h264)")
-        # except Exception:
-        fourcc = cv2.VideoWriter_fourcc(*'mp4v')
-        out_vid = cv2.VideoWriter(vid_output_path, fourcc, fps, (cam_width, cam_height))
-            # logging.info("Failed to open video writer with 'avc1' (h264). Using 'mp4v' instead.")
-
-    return cap, out_vid, cam_width, cam_height, fps
-
-
-def setup_video(video_file_path, save_video, vid_output_path):
-    '''
-    Set up video capture with OpenCV.
-
-    INPUTS:
-    - video_file_path: Path. The path to the video file
-    - save_video: bool. Whether to save the video output
-    - vid_output_path: Path. The path to save the video output
-
-    OUTPUTS:
-    - cap: cv2.VideoCapture. The video capture object
-    - out_vid: cv2.VideoWriter. The video writer object
-    - cam_width: int. The width of the video
-    - cam_height: int. The height of the video
-    - fps: int. The frame rate of the video
-    '''
-    
-    if video_file_path.name == video_file_path.stem:
-        raise ValueError("Please set video_input to 'webcam' or to a video file (with extension) in Config.toml")
-    try:
-        cap = cv2.VideoCapture(video_file_path)
-        if not cap.isOpened():
-            raise
-    except:
-        raise NameError(f"{video_file_path} is not a video. Check video_dir and video_input in your Config.toml file.")
-    
-    cam_width = int(cap.get(cv2.CAP_PROP_FRAME_WIDTH))
-    cam_height = int(cap.get(cv2.CAP_PROP_FRAME_HEIGHT))
-    fps = cap.get(cv2.CAP_PROP_FPS)
-    if fps == 0: fps = 30
-    
-    out_vid = None
-
-    if save_video:
-        # try:
-        #     fourcc = cv2.VideoWriter_fourcc(*'avc1') # =h264. better compression and quality but may fail on some systems
-        #     out_vid = cv2.VideoWriter(vid_output_path, fourcc, fps, (cam_width, cam_height))
-        #     if not out_vid.isOpened():
-        #         raise ValueError("Failed to open video writer with 'avc1' (h264)")
-        # except Exception:
-        fourcc = cv2.VideoWriter_fourcc(*'mp4v')
-        out_vid = cv2.VideoWriter(vid_output_path, fourcc, fps, (cam_width, cam_height))
-            # logging.info("Failed to open video writer with 'avc1' (h264). Using 'mp4v' instead.")
-        
-    return cap, out_vid, cam_width, cam_height, fps
-
-def setup_capture_directories(file_path, output_dir):
-    """
-    Sets up directories for output and prepares for video capture.
-
-    Parameters:
-        file_path (str): Path to the file or 'webcam' for webcam usage.
-        output_dir (str): Base directory to store the output directories and files.
-
-    Returns:
-        dict: A dictionary containing paths for image output, JSON output, and output video.
-    """
-    # Create output directories based on the file path or webcam
-    if file_path == "webcam":
-        current_date = datetime.now().strftime("%Y%m%d_%H%M%S")
-        output_dir_name = f'webcam_{current_date}'
-    else:
-        file_stem = os.path.splitext(os.path.basename(file_path))[0]
-        output_dir_name = f'{file_stem}_Sports2D'
-
-    # Define the full path for the output directory
-    output_dir_full = os.path.abspath(os.path.join(output_dir, output_dir_name))
-    
-    # Create output directories if they do not exist
-    if not os.path.isdir(output_dir_full):
-        os.makedirs(output_dir_full)
-    
-    # Prepare directories for images and JSON outputs
-    img_output_dir = os.path.join(output_dir_full, f'{output_dir_name}_img')
-    json_output_dir = os.path.join(output_dir_full, f'{output_dir_name}_json')
-    if not os.path.isdir(img_output_dir):
-        os.makedirs(img_output_dir)
-    if not os.path.isdir(json_output_dir):
-        os.makedirs(json_output_dir)
-    
-    # Define the path for the output video file
-    output_video_path = os.path.join(output_dir_full, f'{output_dir_name}_pose.mp4')
-
-    return output_dir, output_dir_name, img_output_dir, json_output_dir, output_video_path
-
-def validate_video_file(video_file_path):
-    """
-    Validates if the specified path is a video file.
-    Raises an exception if the file is not a video.
-    """
-    import cv2
-    try:
-        cap = cv2.VideoCapture(video_file_path)
-        if not cap.read()[0]:
-            raise ValueError
-    except:
-        raise NameError(f"{video_file_path} is not a video. Images must be put in one subdirectory per camera.")
-
-def setup_video_capture(video_file_path, webcam_id=None, save_video=False, output_video_path=None, input_size=None, input_frame_range=[]):
-    """
-    Sets up video capture from a webcam or a video file. Optionally saves the output.
-    """
-    import cv2, sys, logging
-    from tqdm import tqdm
-
-    validate_video_file(video_file_path)
-
-    if video_file_path == "webcam":
-        cap, out_vid, cam_width, cam_height, fps = setup_webcam(webcam_id, save_video, output_video_path, input_size)
-        frame_range = [0, sys.maxsize]
-        frame_iterator = range(*frame_range)
-        logging.warning('Webcam input: the framerate may vary. If results are filtered, Sports2D will use the average framerate as input.')
-    else:
-        cap, out_vid, cam_width, cam_height, fps = setup_video(video_file_path, save_video, output_video_path)
-        total_frames = int(cap.get(cv2.CAP_PROP_FRAME_COUNT))
-        if input_frame_range == []:
-            frame_range = [0, total_frames]
-        else:
-            frame_range = input_frame_range
-        frame_iterator = tqdm(range(*frame_range), desc=f'Processing {os.path.basename(video_file_path)}') # use a progress bar
-        start_frame = input_frame_range[0]
-        cap.set(cv2.CAP_PROP_POS_FRAMES, start_frame)
-    
-    return cap, frame_iterator, out_vid, cam_width, cam_height, fps
-
-def display_realtime_results(video_file_path):
-    """
-    Sets up a window for real-time video results display.
-    """
-    import cv2
-
-    cv2.namedWindow(f'Pose Estimation {os.path.basename(video_file_path)}', cv2.WINDOW_NORMAL + cv2.WINDOW_KEEPRATIO)
-    cv2.setWindowProperty(f'Pose Estimation {os.path.basename(video_file_path)}', cv2.WND_PROP_ASPECT_RATIO, cv2.WINDOW_FULLSCREEN)
-
-def finalize_video_processing(frames_processed, total_processing_start_time, output_video_path, fps):
-    """
-    Finalizes processing by freeing resources and adjusting the framerate if necessary.
-    """
-
-    total_processing_time = (datetime.now() - total_processing_start_time).total_seconds()
-    if total_processing_time > 0:
-        actual_framerate = frames_processed / total_processing_time
-        logging.info(f"Rewriting webcam video based on the average framerate {actual_framerate:.2f}.")
-        resample_video(output_video_path, fps, actual_framerate)
-        fps = actual_framerate
-
-    return fps
-
-def read_frame(cap, frame_idx):
-    success, frame = cap.read()
-    if not success:
-        logging.warning(f"Failed to grab frame {frame_idx}.")
-        return None
-    return frame
-
-def display_quit_message(frame, cam_width, cam_height, fontSize, thickness):
-    message = "Press 'q' to quit"
-    position = (cam_width - int(400 * fontSize), cam_height - 20)
-    cv2.putText(
-        frame, message, position,
-        cv2.FONT_HERSHEY_SIMPLEX, fontSize + 0.2,
-        (0, 0, 255), thickness, cv2.LINE_AA
-    )
-
-def track_people(keypoints, scores, multi_person, tracking_mode, prev_keypoints, pose_tracker=None):
-    if multi_person:
-        if tracking_mode == 'rtmlib':
-            keypoints, scores = sort_people_rtmlib(pose_tracker, keypoints, scores)
-        else:
-            if prev_keypoints is None: prev_keypoints = keypoints
-            prev_keypoints, keypoints, scores = sort_people_sports2d(prev_keypoints, keypoints, scores)
-    else:
-        keypoints, scores = np.array([keypoints[0]]), np.array([scores[0]])
-        
-    return keypoints, scores, prev_keypoints
-
-def sort_people_rtmlib(pose_tracker, keypoints, scores):
-    '''
-    Associate persons across frames (RTMLib method)
-
-    INPUTS:
-    - pose_tracker: PoseTracker. The initialized RTMLib pose tracker object
-    - keypoints: array of shape K, L, M with K the number of detected persons,
-    L the number of detected keypoints, M their 2D coordinates
-    - scores: array of shape K, L with K the number of detected persons,
-    L the confidence of detected keypoints
-
-    OUTPUT:
-    - sorted_keypoints: array with reordered persons
-    - sorted_scores: array with reordered scores
-    '''
-    
-    try:
-        desired_size = max(pose_tracker.track_ids_last_frame)+1
-        sorted_keypoints = np.full((desired_size, keypoints.shape[1], 2), np.nan)
-        sorted_keypoints[pose_tracker.track_ids_last_frame] = keypoints[:len(pose_tracker.track_ids_last_frame), :, :]
-        sorted_scores = np.full((desired_size, scores.shape[1]), np.nan)
-        sorted_scores[pose_tracker.track_ids_last_frame] = scores[:len(pose_tracker.track_ids_last_frame), :]
-    except:
-        sorted_keypoints, sorted_scores = keypoints, scores
-
-    return sorted_keypoints, sorted_scores
-
-def sort_people_sports2d(keyptpre, keypt, scores):
-    '''
-    Associate persons across frames (Pose2Sim method)
-    Persons' indices are sometimes swapped when changing frame
-    A person is associated to another in the next frame when they are at a small distance
-    
-    N.B.: Requires min_with_single_indices and euclidian_distance function (see common.py)
-
-    INPUTS:
-    - keyptpre: array of shape K, L, M with K the number of detected persons,
-    L the number of detected keypoints, M their 2D coordinates
-    - keypt: idem keyptpre, for current frame
-    - score: array of shape K, L with K the number of detected persons,
-    L the confidence of detected keypoints
-    
-    OUTPUTS:
-    - sorted_prev_keypoints: array with reordered persons with values of previous frame if current is empty
-    - sorted_keypoints: array with reordered persons
-    - sorted_scores: array with reordered scores
-    '''
-    
-    # Generate possible person correspondences across frames
-    if len(keyptpre) < len(keypt):
-        keyptpre = np.concatenate((keyptpre, np.full((len(keypt)-len(keyptpre), keypt.shape[1], 2), np.nan)))
-    if len(keypt) < len(keyptpre):
-        keypt = np.concatenate((keypt, np.full((len(keyptpre)-len(keypt), keypt.shape[1], 2), np.nan)))
-        scores = np.concatenate((scores, np.full((len(keyptpre)-len(scores), scores.shape[1]), np.nan)))
-    personsIDs_comb = sorted(list(it.product(range(len(keyptpre)), range(len(keypt)))))
-    
-    # Compute distance between persons from one frame to another
-    frame_by_frame_dist = []
-    for comb in personsIDs_comb:
-        frame_by_frame_dist += [euclidean_distance(keyptpre[comb[0]],keypt[comb[1]])]
-    # frame_by_frame_dist = np.mean(frame_by_frame_dist, axis=1)
-    
-    # Sort correspondences by distance
-    _, _, associated_tuples = min_with_single_indices(frame_by_frame_dist, personsIDs_comb)
-    
-    # Associate points to same index across frames, nan if no correspondence
-    sorted_keypoints, sorted_scores = [], []
-    for i in range(len(keyptpre)):
-        id_in_old =  associated_tuples[:,1][associated_tuples[:,0] == i].tolist()
-        if len(id_in_old) > 0:
-            sorted_keypoints += [keypt[id_in_old[0]]]
-            sorted_scores += [scores[id_in_old[0]]]
-        else:
-            sorted_keypoints += [keypt[i]]
-            sorted_scores += [scores[i]]
-    sorted_keypoints, sorted_scores = np.array(sorted_keypoints), np.array(sorted_scores)
-
-    # Keep track of previous values even when missing for more than one frame
-    sorted_prev_keypoints = np.where(np.isnan(sorted_keypoints) & ~np.isnan(keyptpre), keyptpre, sorted_keypoints)
-    
-    return sorted_prev_keypoints, sorted_keypoints, sorted_scores
-
-def min_with_single_indices(L, T):
-    '''
-    Let L be a list (size s) with T associated tuple indices (size s).
-    Select the smallest values of L, considering that 
-    the next smallest value cannot have the same numbers 
-    in the associated tuple as any of the previous ones.
-
-    Example:
-    L = [  20,   27,  51,    33,   43,   23,   37,   24,   4,   68,   84,    3  ]
-    T = list(it.product(range(2),range(3)))
-      = [(0,0),(0,1),(0,2),(0,3),(1,0),(1,1),(1,2),(1,3),(2,0),(2,1),(2,2),(2,3)]
-
-    - 1st smallest value: 3 with tuple (2,3), index 11
-    - 2nd smallest value when excluding indices (2,.) and (.,3), i.e. [(0,0),(0,1),(0,2),X,(1,0),(1,1),(1,2),X,X,X,X,X]:
-    20 with tuple (0,0), index 0
-    - 3rd smallest value when excluding [X,X,X,X,X,(1,1),(1,2),X,X,X,X,X]:
-    23 with tuple (1,1), index 5
-    
-    INPUTS:
-    - L: list (size s)
-    - T: T associated tuple indices (size s)
-
-    OUTPUTS: 
-    - minL: list of smallest values of L, considering constraints on tuple indices
-    - argminL: list of indices of smallest values of L (indices of best combinations)
-    - T_minL: list of tuples associated with smallest values of L
-    '''
-
-    minL = [np.nanmin(L)]
-    argminL = [np.nanargmin(L)]
-    T_minL = [T[argminL[0]]]
-    
-    mask_tokeep = np.array([True for t in T])
-    i=0
-    while mask_tokeep.any()==True:
-        mask_tokeep = mask_tokeep & np.array([t[0]!=T_minL[i][0] and t[1]!=T_minL[i][1] for t in T])
-        if mask_tokeep.any()==True:
-            indicesL_tokeep = np.where(mask_tokeep)[0]
-            minL += [np.nanmin(np.array(L)[indicesL_tokeep]) if not np.isnan(np.array(L)[indicesL_tokeep]).all() else np.nan]
-            argminL += [indicesL_tokeep[np.nanargmin(np.array(L)[indicesL_tokeep])] if not np.isnan(minL[-1]) else indicesL_tokeep[0]]
-            T_minL += (T[argminL[i+1]],)
-            i+=1
-    
-    return np.array(minL), np.array(argminL), np.array(T_minL)
-
-
-def process_coordinates_and_angles(keypoints, scores, keypoint_likelihood_threshold, keypoint_number_threshold, average_likelihood_threshold, flip_left_right, L_R_direction_idx, keypoints_names, keypoints_ids, angle_names, angle_dict):
-    valid_X, valid_Y, valid_scores = [], [], []
-    valid_X_flipped, valid_angles = [], []
-    for kpts, scrs in zip(keypoints, scores):
-        # Filtrage des keypoints avec faible confiance
-        mask = scrs >= keypoint_likelihood_threshold
-        person_X = np.where(mask, kpts[:, 0], np.nan)
-        person_Y = np.where(mask, kpts[:, 1], np.nan)
-        person_scores = np.where(mask, scrs, np.nan)
-
-        # Vérification du nombre de keypoints valides
-        valid_ratio = np.sum(mask) / len(mask)
-        average_score = np.nanmean(person_scores)
-        if valid_ratio < keypoint_number_threshold or average_score < average_likelihood_threshold:
-            person_X[:] = np.nan
-            person_Y[:] = np.nan
-            person_scores[:] = np.nan
-
-        valid_X.append(person_X)
-        valid_Y.append(person_Y)
-        valid_scores.append(person_scores)
-
-        # Gestion du flip gauche-droite
-        if flip_left_right:
-            person_X_flipped = flip_left_right_direction(
-                person_X, L_R_direction_idx, keypoints_names, keypoints_ids
-            )
-        else:
-            person_X_flipped = person_X.copy()
-        valid_X_flipped.append(person_X_flipped)
-
-        # Calcul des angles
-        person_angles = [
-            compute_angle(
-                ang_name, person_X_flipped, person_Y, angle_dict, keypoints_ids, keypoints_names
-            )
-            for ang_name in angle_names
-        ]
-        valid_angles.append(person_angles)
-    return valid_X, valid_Y, valid_X_flipped, valid_angles
-
-def flip_left_right_direction(person_X, L_R_direction_idx, keypoints_names, keypoints_ids):
-    '''
-    Flip the points to the right or left for more consistent angle calculation 
-    depending on which direction the person is facing
-
-    INPUTS:
-    - person_X: list of x coordinates
-    - L_R_direction_idx: list of indices of the left toe, left heel, right toe, right heel
-    - keypoints_names: list of keypoint names (see skeletons.py)
-    - keypoints_ids: list of keypoint ids (see skeletons.py)
-
-    OUTPUTS:
-    - person_X_flipped: list of x coordinates after flipping
-    '''
-
-    Ltoe_idx, LHeel_idx, Rtoe_idx, RHeel_idx = L_R_direction_idx
-    right_orientation = person_X[Rtoe_idx] - person_X[RHeel_idx]
-    left_orientation = person_X[Ltoe_idx] - person_X[LHeel_idx]
-    global_orientation = right_orientation + left_orientation
-    
-    person_X_flipped = person_X.copy()
-    if left_orientation < 0:
-        for k in keypoints_names:
-            if k.startswith('L'):
-                keypt_idx = keypoints_ids[keypoints_names.index(k)]
-                person_X_flipped[keypt_idx] = person_X_flipped[keypt_idx] * -1
-    if right_orientation < 0:
-        for k in keypoints_names:
-            if k.startswith('R'):
-                keypt_idx = keypoints_ids[keypoints_names.index(k)]
-                person_X_flipped[keypt_idx] = person_X_flipped[keypt_idx] * -1
-    if global_orientation < 0:
-        for k in keypoints_names:
-            if not k.startswith('L') and not k.startswith('R'):
-                keypt_idx = keypoints_ids[keypoints_names.index(k)]
-                person_X_flipped[keypt_idx] = person_X_flipped[keypt_idx] * -1
-    
-    return person_X_flipped
-
-def compute_angle(ang_name, person_X_flipped, person_Y, angle_dict, keypoints_ids, keypoints_names):
-    '''
-    Compute the angles from the 2D coordinates of the keypoints.
-    Takes into account which side the participant is facing.
-    Takes into account the offset and scaling of the angle from angle_dict.
-    Requires points2D_to_angles function (see common.py)
-
-    INPUTS:
-    - ang_name: str. The name of the angle to compute
-    - person_X_flipped: list of x coordinates after flipping if needed
-    - person_Y: list of y coordinates
-    - angle_dict: dict. The dictionary of angles to compute (name: [keypoints, type, offset, scaling])
-    - keypoints_ids: list of keypoint ids (see skeletons.py)
-    - keypoints_names: list of keypoint names (see skeletons.py)
-
-    OUTPUTS:
-    - ang: float. The computed angle
-    '''
-
-    ang_params = angle_dict.get(ang_name)
-    if ang_params is not None:
-        if ang_name in ['pelvis', 'trunk', 'shoulders']:
-            angle_coords = [[np.abs(person_X_flipped[keypoints_ids[keypoints_names.index(kpt)]]), person_Y[keypoints_ids[keypoints_names.index(kpt)]]] for kpt in ang_params[0] if kpt in keypoints_names]
-        else:
-            angle_coords = [[person_X_flipped[keypoints_ids[keypoints_names.index(kpt)]], person_Y[keypoints_ids[keypoints_names.index(kpt)]]] for kpt in ang_params[0] if kpt in keypoints_names]
-        ang = points2D_to_angles(angle_coords)
-        ang += ang_params[2]
-        ang *= ang_params[3]
-        if ang_name in ['pelvis', 'shoulders']:
-            ang = ang-180 if ang>90 else ang
-            ang = ang+180 if ang<-90 else ang
-        else:
-            ang = ang-360 if ang>180 else ang
-            ang = ang+360 if ang<-180 else ang
-    else:
-        ang = np.nan
-
-    return ang
-
-def save_output(img_show, save_video, out_vid, save_images, img_output_dir, output_dir_name, frame_idx):
-    if save_video:
-        out_vid.write(img_show)
-    if save_images:
-        os.makedirs(img_output_dir, exist_ok=True)
-        cv2.imwrite(
-            os.path.join(img_output_dir, f'{output_dir_name}_{frame_idx:06d}.jpg'),
-            img_show
-        )
-
-def setup_logging(dir):
-    '''
-    Create logging file and stream handlers
-    '''
-
-    logging.basicConfig(format='%(message)s', level=logging.INFO,
-        handlers = [logging.handlers.TimedRotatingFileHandler(os.path.join(dir, 'logs.txt'), when='D', interval=7), logging.StreamHandler()])
-
-def get_leaf_keys(config, prefix=''):
-    '''
-    Flatten configuration to map leaf keys to their full path
-    '''
-
-    leaf_keys = {}
-    for key, value in config.items():
-        if isinstance(value, dict):
-            leaf_keys.update(get_leaf_keys(value, prefix=prefix + key + '.'))
-        else:
-            leaf_keys[prefix + key] = value
-    return leaf_keys
-
-
-def set_nested_value(config, flat_key, value):
-    '''
-    Update the nested dictionary based on flattened keys
-    '''
-
-    keys = flat_key.split('.')
-    d = config
-    for key in keys[:-1]:
-        d = d.setdefault(key, {})
-    d[keys[-1]] = value
-
-
-def str2bool(v):
-    '''
-    Convert a string to a boolean value.
-    '''
-    
-    if isinstance(v, bool):
-        return v
-    if v.lower() in ('yes', 'true', 't', 'y', '1'):
-        return True
-    elif v.lower() in ('no', 'false', 'f', 'n', '0'):
-        return False
-    else:
-        raise argparse.ArgumentTypeError('Boolean value expected.')
-
