+++ Sports2D/Utilities/video_management.py
@@ -1,0 +1,527 @@
+#!/usr/bin/env python
+# -*- coding: utf-8 -*-
+
+
+'''
+    ##################################################
+    ## Video Management and Tracking                ##
+    ##################################################
+
+    - Displays real-time results during video processing.
+    - Finalizes video processing and saves outputs.
+    - Tracks people or objects in video streams.
+    - Sorts detected people using RTMLib configurations.
+    - Sorts sports figures using 2D coordinates in sports analysis.
+    - Saves output data and results from video processing.
+    - A function for drawing a dotted line on images or videos.
+    - A function for drawing bounding boxes around detected objects.
+    - A function for visualizing skeletons of people or articulated objects.
+    - A function for displaying detected keypoints on images.
+    - A function for visualizing calculated angles directly on the video.
+    - A function for drawing angles between segments.
+    - A function for drawing angles at joints.
+    - A function for writing angle values directly on the body in images.
+    - A function for displaying angles as a list alongside images.
+'''
+
+
+## INIT
+import os
+import cv2
+import logging
+import numpy as np
+import itertools as it
+import matplotlib.pyplot as plt
+
+from datetime import datetime
+from anytree import PreOrderIter
+
+from Sports2D.Utilities.data_processing import resample_video, euclidean_distance, min_with_single_indices
+from Sports2D.process import angle_dict
+
+
+## AUTHORSHIP INFORMATION
+__author__ = "David Pagnon"
+__copyright__ = "Copyright 2023, Sports2D"
+__credits__ = ["David Pagnon"]
+__license__ = "BSD 3-Clause License"
+__version__ = "0.4.0"
+__maintainer__ = "David Pagnon"
+__email__ = "contact@david-pagnon.com"
+__status__ = "Development"
+
+
+## FUNCTIONS
+def display_realtime_results(file_path):
+    """
+    Sets up a window for real-time video results display.
+    """
+    import cv2
+
+    cv2.namedWindow(f'Pose Estimation {os.path.basename(file_path)}', cv2.WINDOW_NORMAL + cv2.WINDOW_KEEPRATIO)
+    cv2.setWindowProperty(f'Pose Estimation {os.path.basename(file_path)}', cv2.WND_PROP_ASPECT_RATIO, cv2.WINDOW_FULLSCREEN)
+
+
+def finalize_video_processing(frames_processed, total_processing_start_time, output_video_path, fps):
+    """
+    Finalizes processing by freeing resources and adjusting the framerate if necessary.
+    """
+
+    total_processing_time = (datetime.now() - total_processing_start_time).total_seconds()
+    if total_processing_time > 0:
+        actual_framerate = frames_processed / total_processing_time
+        logging.info(f"Rewriting webcam video based on the average framerate {actual_framerate:.2f}.")
+        resample_video(output_video_path, fps, actual_framerate)
+        fps = actual_framerate
+
+    return fps
+
+
+def track_people(keypoints, scores, multi_person, tracking_mode, prev_keypoints, pose_tracker=None):
+    if multi_person:
+        if tracking_mode == 'rtmlib':
+            keypoints, scores = sort_people_rtmlib(pose_tracker, keypoints, scores)
+        else:
+            if prev_keypoints is None: prev_keypoints = keypoints
+            prev_keypoints, keypoints, scores = sort_people_sports2d(prev_keypoints, keypoints, scores)
+    else:
+        keypoints, scores = np.array([keypoints[0]]), np.array([scores[0]])
+        
+    return keypoints, scores, prev_keypoints
+
+
+def sort_people_rtmlib(pose_tracker, keypoints, scores):
+    '''
+    Associate persons across frames (RTMLib method)
+
+    INPUTS:
+    - pose_tracker: PoseTracker. The initialized RTMLib pose tracker object
+    - keypoints: array of shape K, L, M with K the number of detected persons,
+    L the number of detected keypoints, M their 2D coordinates
+    - scores: array of shape K, L with K the number of detected persons,
+    L the confidence of detected keypoints
+
+    OUTPUT:
+    - sorted_keypoints: array with reordered persons
+    - sorted_scores: array with reordered scores
+    '''
+    
+    try:
+        desired_size = max(pose_tracker.track_ids_last_frame)+1
+        sorted_keypoints = np.full((desired_size, keypoints.shape[1], 2), np.nan)
+        sorted_keypoints[pose_tracker.track_ids_last_frame] = keypoints[:len(pose_tracker.track_ids_last_frame), :, :]
+        sorted_scores = np.full((desired_size, scores.shape[1]), np.nan)
+        sorted_scores[pose_tracker.track_ids_last_frame] = scores[:len(pose_tracker.track_ids_last_frame), :]
+    except:
+        sorted_keypoints, sorted_scores = keypoints, scores
+
+    return sorted_keypoints, sorted_scores
+
+
+def sort_people_sports2d(keyptpre, keypt, scores):
+    '''
+    Associate persons across frames (Pose2Sim method)
+    Persons' indices are sometimes swapped when changing frame
+    A person is associated to another in the next frame when they are at a small distance
+    
+    N.B.: Requires min_with_single_indices and euclidian_distance function (see common.py)
+
+    INPUTS:
+    - keyptpre: array of shape K, L, M with K the number of detected persons,
+    L the number of detected keypoints, M their 2D coordinates
+    - keypt: idem keyptpre, for current frame
+    - score: array of shape K, L with K the number of detected persons,
+    L the confidence of detected keypoints
+    
+    OUTPUTS:
+    - sorted_prev_keypoints: array with reordered persons with values of previous frame if current is empty
+    - sorted_keypoints: array with reordered persons
+    - sorted_scores: array with reordered scores
+    '''
+    
+    # Generate possible person correspondences across frames
+    if len(keyptpre) < len(keypt):
+        keyptpre = np.concatenate((keyptpre, np.full((len(keypt)-len(keyptpre), keypt.shape[1], 2), np.nan)))
+    if len(keypt) < len(keyptpre):
+        keypt = np.concatenate((keypt, np.full((len(keyptpre)-len(keypt), keypt.shape[1], 2), np.nan)))
+        scores = np.concatenate((scores, np.full((len(keyptpre)-len(scores), scores.shape[1]), np.nan)))
+    personsIDs_comb = sorted(list(it.product(range(len(keyptpre)), range(len(keypt)))))
+    
+    # Compute distance between persons from one frame to another
+    frame_by_frame_dist = []
+    for comb in personsIDs_comb:
+        frame_by_frame_dist += [euclidean_distance(keyptpre[comb[0]],keypt[comb[1]])]
+    # frame_by_frame_dist = np.mean(frame_by_frame_dist, axis=1)
+    
+    # Sort correspondences by distance
+    _, _, associated_tuples = min_with_single_indices(frame_by_frame_dist, personsIDs_comb)
+    
+    # Associate points to same index across frames, nan if no correspondence
+    sorted_keypoints, sorted_scores = [], []
+    for i in range(len(keyptpre)):
+        id_in_old =  associated_tuples[:,1][associated_tuples[:,0] == i].tolist()
+        if len(id_in_old) > 0:
+            sorted_keypoints += [keypt[id_in_old[0]]]
+            sorted_scores += [scores[id_in_old[0]]]
+        else:
+            sorted_keypoints += [keypt[i]]
+            sorted_scores += [scores[i]]
+    sorted_keypoints, sorted_scores = np.array(sorted_keypoints), np.array(sorted_scores)
+
+    # Keep track of previous values even when missing for more than one frame
+    sorted_prev_keypoints = np.where(np.isnan(sorted_keypoints) & ~np.isnan(keyptpre), keyptpre, sorted_keypoints)
+    
+    return sorted_prev_keypoints, sorted_keypoints, sorted_scores
+
+
+def save_output(img_show, save_video, out_vid, save_images, img_output_dir, output_dir_name, frame_idx):
+    if save_video:
+        out_vid.write(img_show)
+    if save_images:
+        os.makedirs(img_output_dir, exist_ok=True)
+        cv2.imwrite(
+            os.path.join(img_output_dir, f'{output_dir_name}_{frame_idx:06d}.jpg'),
+            img_show
+        )
+
+
+def draw_dotted_line(img, start, direction, length, color=(0, 255, 0), gap=7, dot_length=3, thickness=1):
+    '''
+    Draw a dotted line with on a cv2 image
+
+    INPUTS:
+    - img: opencv image
+    - start: np.array. The starting point of the line
+    - direction: np.array. The direction of the line
+    - length: int. The length of the line
+    - color: tuple. The color of the line
+    - gap: int. The distance between each dot
+    - dot_length: int. The length of each dot
+    - thickness: int. The thickness of the line
+
+    OUTPUT:
+    - img: image with the dotted line
+    '''
+
+    for i in range(0, length, gap):
+        line_start = start + direction * i
+        line_end = line_start + direction * dot_length
+        cv2.line(img, tuple(line_start.astype(int)), tuple(line_end.astype(int)), color, thickness)
+
+
+def draw_bounding_box(img, X, Y, colors=[(255, 0, 0), (0, 255, 0), (0, 0, 255)], fontSize=0.3, thickness=1):
+    '''
+    Draw bounding boxes and person ID around list of lists of X and Y coordinates.
+    Bounding boxes have a different color for each person.
+    
+    INPUTS:
+    - img: opencv image
+    - X: list of list of x coordinates
+    - Y: list of list of y coordinates
+    - colors: list of colors to cycle through
+    
+    OUTPUT:
+    - img: image with rectangles and person IDs
+    '''
+   
+    color_cycle = it.cycle(colors)
+
+    for i,(x,y) in enumerate(zip(X,Y)):
+        color = next(color_cycle)
+        if not np.isnan(x).all():
+            x_min, y_min = np.nanmin(x).astype(int), np.nanmin(y).astype(int)
+            x_max, y_max = np.nanmax(x).astype(int), np.nanmax(y).astype(int)
+            if x_min < 0: x_min = 0
+            if x_max > img.shape[1]: x_max = img.shape[1]
+            if y_min < 0: y_min = 0
+            if y_max > img.shape[0]: y_max = img.shape[0]
+
+            # Draw rectangles
+            cv2.rectangle(img, (x_min-25, y_min-25), (x_max+25, y_max+25), color, thickness) 
+        
+            # Write person ID
+            cv2.putText(img, str(i), (x_min-30, y_min-30), cv2.FONT_HERSHEY_SIMPLEX, fontSize+1, color, 2, cv2.LINE_AA) 
+    
+    return img
+
+
+def draw_skel(img, X, Y, model, colors=[(255, 0, 0), (0, 255, 0), (0, 0, 255)], thickness=1):
+    '''
+    Draws keypoints and skeleton for each person.
+    Skeletons have a different color for each person.
+
+    INPUTS:
+    - img: opencv image
+    - X: list of list of x coordinates
+    - Y: list of list of y coordinates
+    - model: skeleton model (from skeletons.py)
+    - colors: list of colors to cycle through
+    
+    OUTPUT:
+    - img: image with keypoints and skeleton
+    '''
+    
+    # Get (unique) pairs between which to draw a line
+    node_pairs = []
+    for data_i in PreOrderIter(model.root, filter_=lambda node: node.is_leaf):
+        node_branches = [node_i.id for node_i in data_i.path]
+        node_pairs += [[node_branches[i],node_branches[i+1]] for i in range(len(node_branches)-1)]
+    node_pairs = [list(x) for x in set(tuple(x) for x in node_pairs)]
+    
+    # Draw lines
+    color_cycle = it.cycle(colors)
+    for (x,y) in zip(X,Y):
+        c = next(color_cycle)
+        if not np.isnan(x).all():
+            [cv2.line(img,
+                (int(x[n[0]]), int(y[n[0]])), (int(x[n[1]]), int(y[n[1]])), c, thickness)
+                for n in node_pairs
+                if not (np.isnan(x[n[0]]) or np.isnan(y[n[0]]) or np.isnan(x[n[1]]) or np.isnan(y[n[1]]))]
+
+    return img
+
+
+def draw_keypts(img, X, Y, scores, thickness=1, cmap_str='RdYlGn'):
+    '''
+    Draws keypoints and skeleton for each person.
+    Keypoints' colors depend on their score.
+
+    INPUTS:
+    - img: opencv image
+    - X: list of list of x coordinates
+    - Y: list of list of y coordinates
+    - scores: list of list of scores
+    - cmap_str: colormap name
+    
+    OUTPUT:
+    - img: image with keypoints and skeleton
+    '''
+    
+    scores = np.where(np.isnan(scores), 0, scores)
+    # scores = (scores - 0.4) / (1-0.4) # to get a red color for scores lower than 0.4
+    scores = np.where(scores>0.99, 0.99, scores)
+    scores = np.where(scores<0, 0, scores)
+    
+    cmap = plt.get_cmap(cmap_str)
+    for (x,y,s) in zip(X,Y,scores):
+        c_k = np.array(cmap(s))[:,:-1]*255
+        [cv2.circle(img, (int(x[i]), int(y[i])), thickness + 4, c_k[i][::-1], -1)
+            for i in range(len(x))
+            if not (np.isnan(x[i]) or np.isnan(y[i]))]
+
+    return img
+
+
+def draw_angles(img, valid_X, valid_Y, valid_angles, valid_X_flipped, keypoints_ids, keypoints_names, angle_names, display_angle_values_on= ['body', 'list'], colors=[(255, 0, 0), (0, 255, 0), (0, 0, 255)], fontSize=0.3, thickness=1):
+    '''
+    Draw angles on the image.
+    Angles are displayed as a list on the image and/or on the body.
+
+    INPUTS:
+    - img: opencv image
+    - valid_X: list of list of x coordinates
+    - valid_Y: list of list of y coordinates
+    - valid_angles: list of list of angles
+    - valid_X_flipped: list of list of x coordinates after flipping if needed
+    - keypoints_ids: list of keypoint ids (see skeletons.py)
+    - keypoints_names: list of keypoint names (see skeletons.py)
+    - angle_names: list of angle names
+    - display_angle_values_on: list of str. 'body' and/or 'list'
+    - colors: list of colors to cycle through
+
+    OUTPUT:
+    - img: image with angles
+    '''
+
+    color_cycle = it.cycle(colors)
+    for person_id, (X,Y,angles, X_flipped) in enumerate(zip(valid_X, valid_Y, valid_angles, valid_X_flipped)):
+        c = next(color_cycle)
+        if not np.isnan(X).all():
+            # person label
+            if 'list' in display_angle_values_on:
+                person_label_position = (int(10 + fontSize*150/0.3*person_id), int(fontSize*50))
+                cv2.putText(img, f'person {person_id}', person_label_position, cv2.FONT_HERSHEY_SIMPLEX, fontSize+0.2, (255,255,255), thickness+1, cv2.LINE_AA)
+                cv2.putText(img, f'person {person_id}', person_label_position, cv2.FONT_HERSHEY_SIMPLEX, fontSize+0.2, c, thickness, cv2.LINE_AA)
+            
+            # angle lines, names and values
+            ang_label_line = 1
+            for k, ang in enumerate(angles):
+                if not np.isnan(ang):
+                    ang_name = angle_names[k]
+                    ang_params = angle_dict.get(ang_name)
+                    if ang_params is not None:
+                        ang_coords = np.array([[X[keypoints_ids[keypoints_names.index(kpt)]], Y[keypoints_ids[keypoints_names.index(kpt)]]] for kpt in ang_params[0] if kpt in keypoints_names])
+                        X_flipped_coords = [X_flipped[keypoints_ids[keypoints_names.index(kpt)]] for kpt in ang_params[0] if kpt in keypoints_names]
+                        flip = -1 if any(x_flipped < 0 for x_flipped in X_flipped_coords) else 1
+                        flip = 1 if ang_name in ['pelvis', 'trunk', 'shoulders'] else flip
+                        right_angle = True if ang_params[2]==90 else False
+                        
+                        # Draw angle
+                        if len(ang_coords) == 2: # segment angle
+                            app_point, vec = draw_segment_angle(img, ang_coords, flip)
+                        else: # joint angle
+                            app_point, vec1, vec2 = draw_joint_angle(img, ang_coords, flip, right_angle)
+    
+                        # Write angle on body
+                        if 'body' in display_angle_values_on:
+                            if len(ang_coords) == 2: # segment angle
+                                write_angle_on_body(img, ang, app_point, vec, np.array([1,0]), dist=20, color=(255,255,255), fontSize=fontSize, thickness=thickness)
+                            else: # joint angle
+                                write_angle_on_body(img, ang, app_point, vec1, vec2, dist=40, color=(0,255,0), fontSize=fontSize, thickness=thickness)
+
+                        # Write angle as a list on image with progress bar
+                        if 'list' in display_angle_values_on:
+                            if len(ang_coords) == 2: # segment angle
+                                ang_label_line = write_angle_as_list(img, ang, ang_name, person_label_position, ang_label_line, color = (255,255,255), fontSize=fontSize, thickness=thickness)
+                            else:
+                                ang_label_line = write_angle_as_list(img, ang, ang_name, person_label_position, ang_label_line, color = (0,255,0), fontSize=fontSize, thickness=thickness)
+
+    return img
+
+
+def draw_segment_angle(img, ang_coords, flip, thickness = 1):
+    '''
+    Draw a segment angle on the image.
+
+    INPUTS:
+    - img: opencv image
+    - ang_coords: np.array. The 2D coordinates of the keypoints
+    - flip: int. Whether the angle should be flipped
+
+    OUTPUT:
+    - app_point: np.array. The point where the angle is displayed
+    - unit_segment_direction: np.array. The unit vector of the segment direction
+    - img: image with the angle
+    '''
+    
+    if not np.any(np.isnan(ang_coords)):
+        app_point = np.int32(np.mean(ang_coords, axis=0))
+
+        # segment line
+        segment_direction = np.int32(ang_coords[0]) - np.int32(ang_coords[1])
+        if (segment_direction==0).all():
+            return app_point, np.array([0,0])
+        unit_segment_direction = segment_direction/np.linalg.norm(segment_direction)
+        cv2.line(img, app_point, np.int32(app_point+unit_segment_direction*20), (255,255,255), thickness)
+
+        # horizontal line
+        cv2.line(img, app_point, (np.int32(app_point[0])+flip*20, np.int32(app_point[1])), (255,255,255), thickness)
+
+        return app_point, unit_segment_direction
+
+
+def draw_joint_angle(img, ang_coords, flip, right_angle, thickness = 1):
+    '''
+    Draw a joint angle on the image.
+
+    INPUTS:
+    - img: opencv image
+    - ang_coords: np.array. The 2D coordinates of the keypoints
+    - flip: int. Whether the angle should be flipped
+    - right_angle: bool. Whether the angle should be offset by 90 degrees
+
+    OUTPUT:
+    - app_point: np.array. The point where the angle is displayed
+    - unit_segment_direction: np.array. The unit vector of the segment direction
+    - unit_parentsegment_direction: np.array. The unit vector of the parent segment direction
+    - img: image with the angle
+    '''
+    
+    if not np.any(np.isnan(ang_coords)):
+        app_point = np.int32(ang_coords[1])
+        
+        segment_direction = np.int32(ang_coords[0] - ang_coords[1])
+        parentsegment_direction = np.int32(ang_coords[-2] - ang_coords[-1])
+        if (segment_direction==0).all() or (parentsegment_direction==0).all():
+            return app_point, np.array([0,0]), np.array([0,0])
+        
+        if right_angle:
+            segment_direction = np.array([-flip*segment_direction[1], flip*segment_direction[0]])
+            segment_direction, parentsegment_direction = parentsegment_direction, segment_direction
+
+        # segment line
+        unit_segment_direction = segment_direction/np.linalg.norm(segment_direction)
+        cv2.line(img, app_point, np.int32(app_point+unit_segment_direction*40), (0,255,0), thickness)
+        
+        # parent segment dotted line
+        unit_parentsegment_direction = parentsegment_direction/np.linalg.norm(parentsegment_direction)
+        draw_dotted_line(img, app_point, unit_parentsegment_direction, 40, color=(0, 255, 0), gap=7, dot_length=3, thickness=thickness)
+
+        # arc
+        start_angle = np.degrees(np.arctan2(unit_segment_direction[1], unit_segment_direction[0]))
+        end_angle = np.degrees(np.arctan2(unit_parentsegment_direction[1], unit_parentsegment_direction[0]))
+        if abs(end_angle - start_angle) > 180:
+            if end_angle > start_angle: start_angle += 360
+            else: end_angle += 360
+        cv2.ellipse(img, app_point, (20, 20), 0, start_angle, end_angle, (0, 255, 0), thickness)
+
+        return app_point, unit_segment_direction, unit_parentsegment_direction
+
+
+def write_angle_on_body(img, ang, app_point, vec1, vec2, dist=40, color=(255,255,255), fontSize=0.3, thickness=1):
+    '''
+    Write the angle on the body.
+
+    INPUTS:
+    - img: opencv image
+    - ang: float. The angle value to display
+    - app_point: np.array. The point where the angle is displayed
+    - vec1: np.array. The unit vector of the first segment
+    - vec2: np.array. The unit vector of the second segment
+    - dist: int. The distance from the origin where to write the angle
+    - color: tuple. The color of the angle
+
+    OUTPUT:
+    - img: image with the angle
+    '''
+
+    vec_sum = vec1 + vec2
+    if (vec_sum == 0.).all():
+        return
+    unit_vec_sum = vec_sum/np.linalg.norm(vec_sum)
+    text_position = np.int32(app_point + unit_vec_sum*dist)
+    cv2.putText(img, f'{ang:.1f}', text_position, cv2.FONT_HERSHEY_SIMPLEX, fontSize, (0,0,0), thickness+1, cv2.LINE_AA)
+    cv2.putText(img, f'{ang:.1f}', text_position, cv2.FONT_HERSHEY_SIMPLEX, fontSize, color, thickness, cv2.LINE_AA)
+
+
+def write_angle_as_list(img, ang, ang_name, person_label_position, ang_label_line, color=(255,255,255), fontSize=0.3, thickness=1):
+    '''
+    Write the angle as a list on the image with a progress bar.
+
+    INPUTS:
+    - img: opencv image
+    - ang: float. The value of the angle to display
+    - ang_name: str. The name of the angle
+    - person_label_position: tuple. The position of the person label
+    - ang_label_line: int. The line where to write the angle
+    - color: tuple. The color of the angle
+
+    OUTPUT:
+    - ang_label_line: int. The updated line where to write the next angle
+    - img: image with the angle
+    '''
+    
+    if not np.any(np.isnan(ang)):
+        # angle names and values
+        ang_label_position = (person_label_position[0], person_label_position[1]+int((ang_label_line)*40*fontSize))
+        ang_value_position = (ang_label_position[0]+int(250*fontSize), ang_label_position[1])
+        cv2.putText(img, f'{ang_name}:', ang_label_position, cv2.FONT_HERSHEY_SIMPLEX, fontSize, (0, 0, 0), thickness+1, cv2.LINE_AA)
+        cv2.putText(img, f'{ang_name}:', ang_label_position, cv2.FONT_HERSHEY_SIMPLEX, fontSize, color, thickness, cv2.LINE_AA)
+        cv2.putText(img, f'{ang:.1f}', ang_value_position, cv2.FONT_HERSHEY_SIMPLEX, fontSize, (0, 0, 0), thickness+1, cv2.LINE_AA)
+        cv2.putText(img, f'{ang:.1f}', ang_value_position, cv2.FONT_HERSHEY_SIMPLEX, fontSize, color, thickness, cv2.LINE_AA)
+        
+        # progress bar
+        ang_percent = int(ang*50/180)
+        y_crop, y_crop_end = ang_value_position[1] - int(35*fontSize), ang_value_position[1]
+        x_crop, x_crop_end = ang_label_position[0]+int(300*fontSize), ang_label_position[0]+int(300*fontSize)+int(ang_percent*fontSize/0.3)
+        if ang_percent < 0:
+            x_crop, x_crop_end = x_crop_end, x_crop
+        img_crop = img[y_crop:y_crop_end, x_crop:x_crop_end]
+        if img_crop.size>0:
+            white_rect = np.ones(img_crop.shape, dtype=np.uint8)*255
+            alpha_rect = cv2.addWeighted(img_crop, 0.6, white_rect, 0.4, 1.0)
+            img[y_crop:y_crop_end, x_crop:x_crop_end] = alpha_rect
+
+        ang_label_line += 1
+    
+    return ang_label_line
