+++ Sports2D/process.py
@@ -67,6 +67,10 @@ from Sports2D.Utilities import filter
 from Sports2D.Utilities.common import *
 from Sports2D.Utilities.skeletons import *
 
+from Sports2D.Utilities.video_management import draw_bounding_box, draw_keypts, draw_skel, draw_angles
+from Sports2D.Utilities.data_processing import make_trc_with_XYZ, make_mot_with_angles
+from Sports2D.Utilities.visualisation import pose_plots, angle_plots
+
 
 ## CONSTANTS
 angle_dict = { # lowercase!
@@ -109,7 +113,6 @@ colors = [(255, 0, 0), (0, 0, 255), (255, 255, 0), (255, 0, 255), (0, 255, 255),
             (125, 0, 0), (0, 125, 0), (0, 0, 125), (125, 125, 0), (125, 0, 125), (0, 125, 125), 
             (255, 125, 125), (125, 255, 125), (125, 125, 255), (255, 255, 125), (255, 125, 255), (125, 255, 255), (125, 125, 125),
             (255, 0, 125), (255, 125, 0), (0, 125, 255), (0, 255, 125), (125, 0, 255), (125, 255, 0), (0, 255, 0)]
-thickness = 1
 
 
 ## AUTHORSHIP INFORMATION
@@ -124,503 +127,6 @@ __status__ = "Development"
 
 
 # FUNCTIONS
-def draw_dotted_line(img, start, direction, length, color=(0, 255, 0), gap=7, dot_length=3, thickness=thickness):
-    '''
-    Draw a dotted line with on a cv2 image
-
-    INPUTS:
-    - img: opencv image
-    - start: np.array. The starting point of the line
-    - direction: np.array. The direction of the line
-    - length: int. The length of the line
-    - color: tuple. The color of the line
-    - gap: int. The distance between each dot
-    - dot_length: int. The length of each dot
-    - thickness: int. The thickness of the line
-
-    OUTPUT:
-    - img: image with the dotted line
-    '''
-
-    for i in range(0, length, gap):
-        line_start = start + direction * i
-        line_end = line_start + direction * dot_length
-        cv2.line(img, tuple(line_start.astype(int)), tuple(line_end.astype(int)), color, thickness)
-
-
-def draw_bounding_box(img, X, Y, colors=[(255, 0, 0), (0, 255, 0), (0, 0, 255)], fontSize=0.3, thickness=1):
-    '''
-    Draw bounding boxes and person ID around list of lists of X and Y coordinates.
-    Bounding boxes have a different color for each person.
-    
-    INPUTS:
-    - img: opencv image
-    - X: list of list of x coordinates
-    - Y: list of list of y coordinates
-    - colors: list of colors to cycle through
-    
-    OUTPUT:
-    - img: image with rectangles and person IDs
-    '''
-   
-    color_cycle = it.cycle(colors)
-
-    for i,(x,y) in enumerate(zip(X,Y)):
-        color = next(color_cycle)
-        if not np.isnan(x).all():
-            x_min, y_min = np.nanmin(x).astype(int), np.nanmin(y).astype(int)
-            x_max, y_max = np.nanmax(x).astype(int), np.nanmax(y).astype(int)
-            if x_min < 0: x_min = 0
-            if x_max > img.shape[1]: x_max = img.shape[1]
-            if y_min < 0: y_min = 0
-            if y_max > img.shape[0]: y_max = img.shape[0]
-
-            # Draw rectangles
-            cv2.rectangle(img, (x_min-25, y_min-25), (x_max+25, y_max+25), color, thickness) 
-        
-            # Write person ID
-            cv2.putText(img, str(i), (x_min-30, y_min-30), cv2.FONT_HERSHEY_SIMPLEX, fontSize+1, color, 2, cv2.LINE_AA) 
-    
-    return img
-
-
-def draw_skel(img, X, Y, model, colors=[(255, 0, 0), (0, 255, 0), (0, 0, 255)]):
-    '''
-    Draws keypoints and skeleton for each person.
-    Skeletons have a different color for each person.
-
-    INPUTS:
-    - img: opencv image
-    - X: list of list of x coordinates
-    - Y: list of list of y coordinates
-    - model: skeleton model (from skeletons.py)
-    - colors: list of colors to cycle through
-    
-    OUTPUT:
-    - img: image with keypoints and skeleton
-    '''
-    
-    # Get (unique) pairs between which to draw a line
-    node_pairs = []
-    for data_i in PreOrderIter(model.root, filter_=lambda node: node.is_leaf):
-        node_branches = [node_i.id for node_i in data_i.path]
-        node_pairs += [[node_branches[i],node_branches[i+1]] for i in range(len(node_branches)-1)]
-    node_pairs = [list(x) for x in set(tuple(x) for x in node_pairs)]
-    
-    # Draw lines
-    color_cycle = it.cycle(colors)
-    for (x,y) in zip(X,Y):
-        c = next(color_cycle)
-        if not np.isnan(x).all():
-            [cv2.line(img,
-                (int(x[n[0]]), int(y[n[0]])), (int(x[n[1]]), int(y[n[1]])), c, thickness)
-                for n in node_pairs
-                if not (np.isnan(x[n[0]]) or np.isnan(y[n[0]]) or np.isnan(x[n[1]]) or np.isnan(y[n[1]]))]
-
-    return img
-
-
-def draw_keypts(img, X, Y, scores, cmap_str='RdYlGn'):
-    '''
-    Draws keypoints and skeleton for each person.
-    Keypoints' colors depend on their score.
-
-    INPUTS:
-    - img: opencv image
-    - X: list of list of x coordinates
-    - Y: list of list of y coordinates
-    - scores: list of list of scores
-    - cmap_str: colormap name
-    
-    OUTPUT:
-    - img: image with keypoints and skeleton
-    '''
-    
-    scores = np.where(np.isnan(scores), 0, scores)
-    # scores = (scores - 0.4) / (1-0.4) # to get a red color for scores lower than 0.4
-    scores = np.where(scores>0.99, 0.99, scores)
-    scores = np.where(scores<0, 0, scores)
-    
-    cmap = plt.get_cmap(cmap_str)
-    for (x,y,s) in zip(X,Y,scores):
-        c_k = np.array(cmap(s))[:,:-1]*255
-        [cv2.circle(img, (int(x[i]), int(y[i])), thickness+4, c_k[i][::-1], -1)
-            for i in range(len(x))
-            if not (np.isnan(x[i]) or np.isnan(y[i]))]
-
-    return img
-
-
-def draw_angles(img, valid_X, valid_Y, valid_angles, valid_X_flipped, keypoints_ids, keypoints_names, angle_names, display_angle_values_on= ['body', 'list'], colors=[(255, 0, 0), (0, 255, 0), (0, 0, 255)], fontSize=0.3, thickness=1):
-    '''
-    Draw angles on the image.
-    Angles are displayed as a list on the image and/or on the body.
-
-    INPUTS:
-    - img: opencv image
-    - valid_X: list of list of x coordinates
-    - valid_Y: list of list of y coordinates
-    - valid_angles: list of list of angles
-    - valid_X_flipped: list of list of x coordinates after flipping if needed
-    - keypoints_ids: list of keypoint ids (see skeletons.py)
-    - keypoints_names: list of keypoint names (see skeletons.py)
-    - angle_names: list of angle names
-    - display_angle_values_on: list of str. 'body' and/or 'list'
-    - colors: list of colors to cycle through
-
-    OUTPUT:
-    - img: image with angles
-    '''
-
-    color_cycle = it.cycle(colors)
-    for person_id, (X,Y,angles, X_flipped) in enumerate(zip(valid_X, valid_Y, valid_angles, valid_X_flipped)):
-        c = next(color_cycle)
-        if not np.isnan(X).all():
-            # person label
-            if 'list' in display_angle_values_on:
-                person_label_position = (int(10 + fontSize*150/0.3*person_id), int(fontSize*50))
-                cv2.putText(img, f'person {person_id}', person_label_position, cv2.FONT_HERSHEY_SIMPLEX, fontSize+0.2, (255,255,255), thickness+1, cv2.LINE_AA)
-                cv2.putText(img, f'person {person_id}', person_label_position, cv2.FONT_HERSHEY_SIMPLEX, fontSize+0.2, c, thickness, cv2.LINE_AA)
-            
-            # angle lines, names and values
-            ang_label_line = 1
-            for k, ang in enumerate(angles):
-                if not np.isnan(ang):
-                    ang_name = angle_names[k]
-                    ang_params = angle_dict.get(ang_name)
-                    if ang_params is not None:
-                        ang_coords = np.array([[X[keypoints_ids[keypoints_names.index(kpt)]], Y[keypoints_ids[keypoints_names.index(kpt)]]] for kpt in ang_params[0] if kpt in keypoints_names])
-                        X_flipped_coords = [X_flipped[keypoints_ids[keypoints_names.index(kpt)]] for kpt in ang_params[0] if kpt in keypoints_names]
-                        flip = -1 if any(x_flipped < 0 for x_flipped in X_flipped_coords) else 1
-                        flip = 1 if ang_name in ['pelvis', 'trunk', 'shoulders'] else flip
-                        right_angle = True if ang_params[2]==90 else False
-                        
-                        # Draw angle
-                        if len(ang_coords) == 2: # segment angle
-                            app_point, vec = draw_segment_angle(img, ang_coords, flip)
-                        else: # joint angle
-                            app_point, vec1, vec2 = draw_joint_angle(img, ang_coords, flip, right_angle)
-    
-                        # Write angle on body
-                        if 'body' in display_angle_values_on:
-                            if len(ang_coords) == 2: # segment angle
-                                write_angle_on_body(img, ang, app_point, vec, np.array([1,0]), dist=20, color=(255,255,255), fontSize=fontSize, thickness=thickness)
-                            else: # joint angle
-                                write_angle_on_body(img, ang, app_point, vec1, vec2, dist=40, color=(0,255,0), fontSize=fontSize, thickness=thickness)
-
-                        # Write angle as a list on image with progress bar
-                        if 'list' in display_angle_values_on:
-                            if len(ang_coords) == 2: # segment angle
-                                ang_label_line = write_angle_as_list(img, ang, ang_name, person_label_position, ang_label_line, color = (255,255,255), fontSize=fontSize, thickness=thickness)
-                            else:
-                                ang_label_line = write_angle_as_list(img, ang, ang_name, person_label_position, ang_label_line, color = (0,255,0), fontSize=fontSize, thickness=thickness)
-
-    return img
-
-
-def draw_segment_angle(img, ang_coords, flip):
-    '''
-    Draw a segment angle on the image.
-
-    INPUTS:
-    - img: opencv image
-    - ang_coords: np.array. The 2D coordinates of the keypoints
-    - flip: int. Whether the angle should be flipped
-
-    OUTPUT:
-    - app_point: np.array. The point where the angle is displayed
-    - unit_segment_direction: np.array. The unit vector of the segment direction
-    - img: image with the angle
-    '''
-    
-    if not np.any(np.isnan(ang_coords)):
-        app_point = np.int32(np.mean(ang_coords, axis=0))
-
-        # segment line
-        segment_direction = np.int32(ang_coords[0]) - np.int32(ang_coords[1])
-        if (segment_direction==0).all():
-            return app_point, np.array([0,0])
-        unit_segment_direction = segment_direction/np.linalg.norm(segment_direction)
-        cv2.line(img, app_point, np.int32(app_point+unit_segment_direction*20), (255,255,255), thickness)
-
-        # horizontal line
-        cv2.line(img, app_point, (np.int32(app_point[0])+flip*20, np.int32(app_point[1])), (255,255,255), thickness)
-
-        return app_point, unit_segment_direction
-
-
-def draw_joint_angle(img, ang_coords, flip, right_angle):
-    '''
-    Draw a joint angle on the image.
-
-    INPUTS:
-    - img: opencv image
-    - ang_coords: np.array. The 2D coordinates of the keypoints
-    - flip: int. Whether the angle should be flipped
-    - right_angle: bool. Whether the angle should be offset by 90 degrees
-
-    OUTPUT:
-    - app_point: np.array. The point where the angle is displayed
-    - unit_segment_direction: np.array. The unit vector of the segment direction
-    - unit_parentsegment_direction: np.array. The unit vector of the parent segment direction
-    - img: image with the angle
-    '''
-    
-    if not np.any(np.isnan(ang_coords)):
-        app_point = np.int32(ang_coords[1])
-        
-        segment_direction = np.int32(ang_coords[0] - ang_coords[1])
-        parentsegment_direction = np.int32(ang_coords[-2] - ang_coords[-1])
-        if (segment_direction==0).all() or (parentsegment_direction==0).all():
-            return app_point, np.array([0,0]), np.array([0,0])
-        
-        if right_angle:
-            segment_direction = np.array([-flip*segment_direction[1], flip*segment_direction[0]])
-            segment_direction, parentsegment_direction = parentsegment_direction, segment_direction
-
-        # segment line
-        unit_segment_direction = segment_direction/np.linalg.norm(segment_direction)
-        cv2.line(img, app_point, np.int32(app_point+unit_segment_direction*40), (0,255,0), thickness)
-        
-        # parent segment dotted line
-        unit_parentsegment_direction = parentsegment_direction/np.linalg.norm(parentsegment_direction)
-        draw_dotted_line(img, app_point, unit_parentsegment_direction, 40, color=(0, 255, 0), gap=7, dot_length=3, thickness=thickness)
-
-        # arc
-        start_angle = np.degrees(np.arctan2(unit_segment_direction[1], unit_segment_direction[0]))
-        end_angle = np.degrees(np.arctan2(unit_parentsegment_direction[1], unit_parentsegment_direction[0]))
-        if abs(end_angle - start_angle) > 180:
-            if end_angle > start_angle: start_angle += 360
-            else: end_angle += 360
-        cv2.ellipse(img, app_point, (20, 20), 0, start_angle, end_angle, (0, 255, 0), thickness)
-
-        return app_point, unit_segment_direction, unit_parentsegment_direction
-
-
-def write_angle_on_body(img, ang, app_point, vec1, vec2, dist=40, color=(255,255,255), fontSize=0.3, thickness=1):
-    '''
-    Write the angle on the body.
-
-    INPUTS:
-    - img: opencv image
-    - ang: float. The angle value to display
-    - app_point: np.array. The point where the angle is displayed
-    - vec1: np.array. The unit vector of the first segment
-    - vec2: np.array. The unit vector of the second segment
-    - dist: int. The distance from the origin where to write the angle
-    - color: tuple. The color of the angle
-
-    OUTPUT:
-    - img: image with the angle
-    '''
-
-    vec_sum = vec1 + vec2
-    if (vec_sum == 0.).all():
-        return
-    unit_vec_sum = vec_sum/np.linalg.norm(vec_sum)
-    text_position = np.int32(app_point + unit_vec_sum*dist)
-    cv2.putText(img, f'{ang:.1f}', text_position, cv2.FONT_HERSHEY_SIMPLEX, fontSize, (0,0,0), thickness+1, cv2.LINE_AA)
-    cv2.putText(img, f'{ang:.1f}', text_position, cv2.FONT_HERSHEY_SIMPLEX, fontSize, color, thickness, cv2.LINE_AA)
-
-
-def write_angle_as_list(img, ang, ang_name, person_label_position, ang_label_line, color=(255,255,255), fontSize=0.3, thickness=1):
-    '''
-    Write the angle as a list on the image with a progress bar.
-
-    INPUTS:
-    - img: opencv image
-    - ang: float. The value of the angle to display
-    - ang_name: str. The name of the angle
-    - person_label_position: tuple. The position of the person label
-    - ang_label_line: int. The line where to write the angle
-    - color: tuple. The color of the angle
-
-    OUTPUT:
-    - ang_label_line: int. The updated line where to write the next angle
-    - img: image with the angle
-    '''
-    
-    if not np.any(np.isnan(ang)):
-        # angle names and values
-        ang_label_position = (person_label_position[0], person_label_position[1]+int((ang_label_line)*40*fontSize))
-        ang_value_position = (ang_label_position[0]+int(250*fontSize), ang_label_position[1])
-        cv2.putText(img, f'{ang_name}:', ang_label_position, cv2.FONT_HERSHEY_SIMPLEX, fontSize, (0, 0, 0), thickness+1, cv2.LINE_AA)
-        cv2.putText(img, f'{ang_name}:', ang_label_position, cv2.FONT_HERSHEY_SIMPLEX, fontSize, color, thickness, cv2.LINE_AA)
-        cv2.putText(img, f'{ang:.1f}', ang_value_position, cv2.FONT_HERSHEY_SIMPLEX, fontSize, (0, 0, 0), thickness+1, cv2.LINE_AA)
-        cv2.putText(img, f'{ang:.1f}', ang_value_position, cv2.FONT_HERSHEY_SIMPLEX, fontSize, color, thickness, cv2.LINE_AA)
-        
-        # progress bar
-        ang_percent = int(ang*50/180)
-        y_crop, y_crop_end = ang_value_position[1] - int(35*fontSize), ang_value_position[1]
-        x_crop, x_crop_end = ang_label_position[0]+int(300*fontSize), ang_label_position[0]+int(300*fontSize)+int(ang_percent*fontSize/0.3)
-        if ang_percent < 0:
-            x_crop, x_crop_end = x_crop_end, x_crop
-        img_crop = img[y_crop:y_crop_end, x_crop:x_crop_end]
-        if img_crop.size>0:
-            white_rect = np.ones(img_crop.shape, dtype=np.uint8)*255
-            alpha_rect = cv2.addWeighted(img_crop, 0.6, white_rect, 0.4, 1.0)
-            img[y_crop:y_crop_end, x_crop:x_crop_end] = alpha_rect
-
-        ang_label_line += 1
-    
-    return ang_label_line
-
-
-def make_trc_with_XYZ(X, Y, Z, time, trc_path):
-    '''
-    Write a trc file from 3D coordinates and time, compatible with OpenSim.
-
-    INPUTS:
-    - X: pd.DataFrame. The x coordinates of the keypoints
-    - Y: pd.DataFrame. The y coordinates of the keypoints
-    - Z: pd.DataFrame. The z coordinates of the keypoints
-    - time: pd.Series. The time series for the coordinates
-    - trc_path: str. The path where to save the trc file
-
-    OUTPUT:
-    - trc_data: pd.DataFrame. The data that has been written to the TRC file
-    '''
-
-    #Header
-    frame_rate = (len(X)-1)/(time.iloc[-1] - time.iloc[0])
-    DataRate = CameraRate = OrigDataRate = frame_rate
-    NumFrames = len(X)
-    NumMarkers = len(X.columns)
-    keypoint_names = X.columns
-    header_trc = ['PathFileType\t4\t(X/Y/Z)\t' + trc_path, 
-            'DataRate\tCameraRate\tNumFrames\tNumMarkers\tUnits\tOrigDataRate\tOrigDataStartFrame\tOrigNumFrames', 
-            '\t'.join(map(str,[DataRate, CameraRate, NumFrames, NumMarkers, 'm', OrigDataRate, 0, NumFrames])),
-            'Frame#\tTime\t' + '\t\t\t'.join(keypoint_names) + '\t\t',
-            '\t\t'+'\t'.join([f'X{i+1}\tY{i+1}\tZ{i+1}' for i in range(len(keypoint_names))])]
-    
-    # Data
-    trc_data = pd.concat([pd.concat([X.iloc[:,kpt], Y.iloc[:,kpt], Z.iloc[:,kpt]], axis=1) for kpt in range(len(X.columns))], axis=1)
-    trc_data.insert(0, 't', time)
-
-    # Write file
-    with open(trc_path, 'w') as trc_o:
-        [trc_o.write(line+'\n') for line in header_trc]
-        trc_data.to_csv(trc_o, sep='\t', index=True, header=None, lineterminator='\n')
-
-    return trc_data
-
-
-def make_mot_with_angles(angles, time, mot_path):
-    '''
-    Write a mot file from angles and time, compatible with OpenSim.
-
-    INPUTS:
-    - angles: pd.DataFrame. The angles to write
-    - time: pd.Series. The time series for the angles
-    - mot_path: str. The path where to save the mot file
-
-    OUTPUT:
-    - angles: pd.DataFrame. The data that has been written to the MOT file
-    '''
-
-    # Header
-    nRows, nColumns = angles.shape
-    angle_names = angles.columns
-    header_mot = ['Coordinates', 
-                  'version=1', 
-                  f'{nRows=}',
-                  f'{nColumns=}',
-                  'inDegrees=yes',
-                  '',
-                  'Units are S.I. units (second, meters, Newtons, ...)',
-                  "If the header above contains a line with 'inDegrees', this indicates whether rotational values are in degrees (yes) or radians (no).",
-                  '',
-                  'endheader',
-                  'time\t' + '\t'.join(angle_names)]
-                  
-    # Write file
-    angles.insert(0,'time',time)
-    with open(mot_path, 'w') as mot_o:
-        [mot_o.write(line+'\n') for line in header_mot]
-        angles.to_csv(mot_o, sep='\t', index=False, header=None, lineterminator='\n')
-
-    return angles
-
-
-def pose_plots(trc_data_unfiltered, trc_data, person_id):
-    '''
-    Displays trc filtered and unfiltered data for comparison
-    /!\ Often crashes on the third window...
-
-    INPUTS:
-    - trc_data_unfiltered: pd.DataFrame. The unfiltered trc data
-    - trc_data: pd.DataFrame. The filtered trc data
-
-    OUTPUT:
-    - matplotlib window with tabbed figures for each keypoint
-    '''
-    
-    mpl.use('qt5agg')
-    mpl.rc('figure', max_open_warning=0)
-
-    keypoints_names = trc_data.columns[1::3]
-    
-    pw = plotWindow()
-    pw.MainWindow.setWindowTitle('Person'+ str(person_id) + ' coordinates') # Main title
-
-    for id, keypoint in enumerate(keypoints_names):
-        f = plt.figure()
-        f.canvas.manager.window.setWindowTitle(keypoint + ' Plot')
-
-        axX = plt.subplot(211)
-        plt.plot(trc_data_unfiltered.iloc[:,0], trc_data_unfiltered.iloc[:,id*3+1], label='unfiltered')
-        plt.plot(trc_data.iloc[:,0], trc_data.iloc[:,id*3+1], label='filtered')
-        plt.setp(axX.get_xticklabels(), visible=False)
-        axX.set_ylabel(keypoint+' X')
-        plt.legend()
-
-        axY = plt.subplot(212)
-        plt.plot(trc_data_unfiltered.iloc[:,0], trc_data_unfiltered.iloc[:,id*3+2], label='unfiltered')
-        plt.plot(trc_data.iloc[:,0], trc_data.iloc[:,id*3+2], label='filtered')
-        axY.set_xlabel('Time (seconds)')
-        axY.set_ylabel(keypoint+' Y')
-
-        pw.addPlot(keypoint, f)
-    
-    pw.show()
-
-
-def angle_plots(angle_data_unfiltered, angle_data, person_id):
-    '''
-    Displays angle filtered and unfiltered data for comparison
-    /!\ Often crashes on the third window...
-
-    INPUTS:
-    - angle_data_unfiltered: pd.DataFrame. The unfiltered angle data
-    - angle_data: pd.DataFrame. The filtered angle data
-
-    OUTPUT:
-    - matplotlib window with tabbed figures for each angle
-    '''
-
-    mpl.use('qt5agg')
-    mpl.rc('figure', max_open_warning=0)
-
-    angles_names = angle_data.columns[1:]
-
-    pw = plotWindow()
-    pw.MainWindow.setWindowTitle('Person'+ str(person_id) + ' angles') # Main title
-
-    for id, angle in enumerate(angles_names):
-        f = plt.figure()
-        
-        ax = plt.subplot(111)
-        plt.plot(angle_data_unfiltered.iloc[:,0], angle_data_unfiltered.iloc[:,id+1], label='unfiltered')
-        plt.plot(angle_data.iloc[:,0], angle_data.iloc[:,id+1], label='filtered')
-        ax.set_xlabel('Time (seconds)')
-        ax.set_ylabel(angle+' (°)')
-        plt.legend()
-
-        pw.addPlot(angle, f)
-
-    pw.show()
-
-
 def process_fun(config_dict, video_file_path, pose_tracker, input_frame_range, output_dir):
     '''
     Detect 2D joint centers from a video or a webcam with RTMLib.
@@ -732,8 +238,9 @@ def process_fun(config_dict, video_file_path, pose_tracker, input_frame_range, o
     frames_processed = 0
     prev_keypoints = None
     for frame_idx in frame_iterator:
-        # If frame not grabbed
         frame = read_frame(cap, frame_idx)
+
+        # If frame not grabbed
         if frame is None:
             logging.warning(f"Failed to grab frame {frame_idx}.")
             if save_pose:
@@ -769,10 +276,10 @@ def process_fun(config_dict, video_file_path, pose_tracker, input_frame_range, o
         # Draw keypoints and skeleton
         if show_realtime_results or save_video or save_images:
             img_show = frame.copy()
-            img_show = draw_bounding_box(img_show, valid_X, valid_Y, colors=colors, fontSize=fontSize, thickness=thickness)
-            img_show = draw_keypts(img_show, valid_X, valid_Y, scores, cmap_str='RdYlGn')
-            img_show = draw_skel(img_show, valid_X, valid_Y, model, colors=colors)
-            img_show = draw_angles(img_show, valid_X, valid_Y, valid_angles, valid_X_flipped, keypoints_ids, keypoints_names, angle_names, display_angle_values_on=display_angle_values_on, colors=colors, fontSize=fontSize, thickness=thickness)
+            img_show = draw_bounding_box(img_show, valid_X, valid_Y, colors, fontSize, thickness)
+            img_show = draw_keypts(img_show, valid_X, valid_Y, scores, thickness, cmap_str='RdYlGn')
+            img_show = draw_skel(img_show, valid_X, valid_Y, model, colors, thickness)
+            img_show = draw_angles(img_show, valid_X, valid_Y, valid_angles, valid_X_flipped, keypoints_ids, keypoints_names, angle_names, display_angle_values_on, colors, fontSize, thickness)
 
         if show_realtime_results:
             cv2.imshow(f"Pose Estimation {os.path.basename(video_file_path)}", img_show)
